Implemented Tasklist/Task SearchBar
The search bar allows for concurrent search of both tasklists and tasks,
which are grouped separately. Tasks entries will include the titles of
their parent tasklists.

Upon clicking on the task/tasklist, the user will be redirected to the
corresponding tasklist (if a task is clicked, then to its parent tasklist)